Massive overhaul (VM support, i3, cleanup)
Added new host (goral) that runs on a VM (Apple Silicon), which supports
VMWare tools.

Added basic i3 support, with alacritty as terminal emulator. This is due
to dwm not running smoothly inside the VM.

Added the hsys.isVM setting, which will set up the system so it supports
running as a guest. And avoids installing packages or adding system
configurations intended for systems running on real hardware.

Added hsys.hidpi setting, which will set up the display for computers
with HiDPI monitors.

Removed following:
- Mate desktop environment & configurations
- Gnome desktop environment & configurations
- Emacs and configurations
- Rofi
- Virtualisations module
- Unused applications
- Backups module